320 - Move types
