+++ lib/types.ts
@@ -7,15 +7,6 @@ export type UserGameInsert = InferInsertModel<typeof UserGamesTable>;
 export type UserGameSelect = InferSelectModel<typeof UserGamesTable>;
 export type UserSelect = InferSelectModel<typeof UsersTable>;
 
-export type UserGameProps = {
-    scenario: ScenarioSelect;
-    remainingScenarios?: number;
-    backstageAccess?: boolean;
-    countdownRunning?: boolean;
-    revealSolution?: boolean;
-    demoMode?: boolean;
-};
-
 export type ScenarioUserGame = Omit<ScenarioSelect, 'data'> & { data: Scenario };
 
 export type CompleteDemoPayload = {
